Add Build Solutions method
The Build Solutions method is the core of the algorithm, it evaluates
and selects the right items to add them to a tabu list to build
solutions.

At this point, the algorithm can give an optimal solution with no
restrictions to select items, it's functional, but those missing
restrictions are strictly necessary for an Knapsack Problem correct
solution.