Include item restriction
The Get Items Prob method fills an int array with a negative one to
initialize it, but when index passes through each item slot on the
probability array calculates the probability of choose that referenced
item and saves it to that index on the probability array.

Whit restrictions, the Get Items Prob method only sets a value on a
probability array slot where it's possible, it means that the item
doesn't exceed the knapsack weight limit and it was no included into the
knapsack before.