Fix errors in Build Solutions method
These changes are included:

General - the prob array now conserves a negative number when an item is
not selectable, so all methods need to consider it.

- Add a break point into loop because it needs to stop when has no items
that can be selected to add them inside knapsack.
- Select Item method sets a value into probability array slot when it's
possible to add that related item, additionally returns a negative
number when has no option to choose anyone.
- Fix Get Random Item, it returned a wrong index of item.