Add Init Ant no capable items restriction
The problem was if the user inputted all items with a weight over the
knapsack max weight limit, so no item was able of being included to
knapsack, because selection of first element didn't stop until find the
first capable item, this caused an endless loop.

Note: this is a new file because it was changed of package.