Add Get Best Solution function
There was no computed way to know what was the best solution, user can
sum the utility and value on each solution and get by himself the best
one.

Now the class includes the Print Best Solution method as way to show
what is the best solution and its properties sumatory after it was
gotten.